restructured project for better separation of concerns
