+++ src/models/board.ts
