+++ src/models/dice.ts
